renamed and moved the default dict to utils
